Tweak aggregator to handle multiple readings, added reader to arduino code

+++ pi/globals.py
@@ -16,7 +16,7 @@ SHMEM_NAME = "mem123"
 # length of sensor data array stored in shm
 SHMEM_NMEM = 10
 # datatype of sensor readings to be stored in shm
-SHMEM_DTYPE = np.float64
+SHMEM_DTYPE = np.int64
 SHMEM_MEMB_SIZE = np.dtype(SHMEM_DTYPE).itemsize
 SHMEM_TOTAL_SIZE = SHMEM_NMEM * SHMEM_MEMB_SIZE
